MCR-3137 move checks to helper class to make manager much more concise
