+++ mycore-user2/src/main/java/org/mycore/user2/hash/MCRPasswordCheckManager.java
@@ -18,20 +18,11 @@
 
 package org.mycore.user2.hash;
 
-import java.io.BufferedReader;
-import java.io.BufferedWriter;
-import java.io.File;
-import java.io.FileReader;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
 import java.security.NoSuchAlgorithmException;
 import java.security.SecureRandom;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
-import java.util.Optional;
 import java.util.Set;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
@@ -39,15 +30,13 @@ import java.util.stream.Collectors;
 import org.mycore.common.MCRException;
 import org.mycore.common.annotation.MCROutdated;
 import org.mycore.common.config.MCRConfiguration2;
-import org.mycore.common.config.MCRConfigurationDir;
-import org.mycore.common.config.MCRConfigurationException;
 import org.mycore.common.config.annotation.MCRConfigurationProxy;
 import org.mycore.common.config.annotation.MCRInstanceMap;
 import org.mycore.common.config.annotation.MCRProperty;
-import org.mycore.resource.MCRResourceHelper;
 
-import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.mycore.common.config.MCRConfiguration2.splitValue;
+import static org.mycore.user2.hash.MCRPasswordCheckManagerHelper.checkConfigurationHasNoIncompatibleChange;
+import static org.mycore.user2.hash.MCRPasswordCheckManagerHelper.checkSelectedStrategyIsNotOutdated;
 
 /**
  * A {@link MCRPasswordCheckManager} can be used to create password hashes and to verify an existing hash
@@ -70,9 +59,10 @@ import static org.mycore.common.config.MCRConfiguration2.splitValue;
  * <li> The selected strategy is configured using the property suffix
  * {@link MCRPasswordCheckManager#SELECTED_STRATEGY_KEY}.
  * <li> The property suffix {@link MCRPasswordCheckManager#CONFIGURATION_CHECKS_KEY} can be used to enable or
- * disable configuration checks during instantiation, specifically: (1) whether the configuration of a selector has
- * been changed in a way that will prevent existing password hashes from being successfully verified, even if the
- * correct password was supplied and (2) whether a strategy annotated with {@link MCROutdated} has been selected.
+ * disable configuration checks during instantiation, specifically: (1) whether a strategy annotated with
+ * {@link MCROutdated} has been selected and (2) whether the configuration of a strategy has been changed in a way
+ * that will prevent existing password hashes from being successfully verified, even if the correct password
+ * was supplied.
  * </ul>
  * Example:
  * <pre>
@@ -92,8 +82,6 @@ public final class MCRPasswordCheckManager {
 
     private static final MCRPasswordCheckManager INSTANCE = instantiate();
 
-    private static final String DATA_DIRECTORY_NAME = "passwordCheckStrategies";
-
     public static final String MANAGER_PROPERTY = "MCR.User.PasswordCheck";
 
     public static final String STRATEGIES_KEY = "Strategies";
@@ -111,7 +99,7 @@ public final class MCRPasswordCheckManager {
     private final String selectedStrategyType;
 
     public MCRPasswordCheckManager(SecureRandom random, Map<String, MCRPasswordCheckStrategy> strategies,
-                                   String selectedStrategyType, Set<ConfigurationCheck> configurationChecks) {
+        String selectedStrategyType, Set<ConfigurationCheck> configurationChecks) {
         this.random = Objects.requireNonNull(random, "Random must not be null");
         this.strategies = new HashMap<>(Objects.requireNonNull(strategies, "Strategies must not be null"));
         this.strategies.values().forEach(strategy -> Objects.requireNonNull(strategy, "Strategy must not be null"));
@@ -127,125 +115,6 @@ public final class MCRPasswordCheckManager {
         if (configurationChecks.contains(ConfigurationCheck.INCOMPATIBLE_CHANGE)) {
             checkConfigurationHasNoIncompatibleChange(strategies);
         }
-    }
-
-    private void checkSelectedStrategyIsNotOutdated(String type, MCRPasswordCheckStrategy strategy) {
-
-        Class<? extends MCRPasswordCheckStrategy> selectedStrategyClass = strategy.getClass();
-
-        if (selectedStrategyClass.isAnnotationPresent(MCROutdated.class)) {
-            throw new MCRConfigurationException("Detected outdated password check strategy implementation " +
-                selectedStrategyClass.getName() + " for selected password check strategy " + type + ", expected " +
-                "an implementation that is not outdated");
-        }
-
-    }
-
-    private void checkConfigurationHasNoIncompatibleChange(Map<String, MCRPasswordCheckStrategy> strategies) {
-
-        for (Map.Entry<String, MCRPasswordCheckStrategy> entry : strategies.entrySet()) {
-
-            String type = entry.getKey();
-            Optional<UnmodifiableConfiguration> oldConfiguration = loadUnmodifiableConfiguration(type);
-            UnmodifiableConfiguration newConfiguration = toUnmodifiableConfiguration(entry.getValue());
-
-            if (oldConfiguration.isPresent()) {
-                checkConfigurationHasNoIncompatibleChange(type, oldConfiguration.get(), newConfiguration);
-            } else {
-                storeUnmodifiableConfiguration(type, newConfiguration);
-            }
-
-        }
-
-    }
-
-    private static Optional<UnmodifiableConfiguration> loadUnmodifiableConfiguration(String type) {
-
-        String path = DATA_DIRECTORY_NAME + "/" + type;
-
-        // previously seen configuration values (those, that must not change)
-        // might have been stored for comparison as a file when a custom provider was first encountered, ...
-        File file = MCRConfigurationDir.getConfigFile(path);
-        if (file != null && file.exists()) {
-            if (!file.isFile() || !file.canRead()) {
-                throw new MCRException("Expected " + file.getAbsolutePath() + " to be a readable file");
-            }
-            try (BufferedReader reader = new BufferedReader(new FileReader(file, UTF_8), 128)) {
-                return Optional.of(readUnmodifiableConfiguration(reader));
-            } catch (IOException e) {
-                throw new MCRException("Unable to read file " + file.getAbsolutePath());
-            }
-        }
-
-        // ... or are provided as a resource in case of a built-in provider
-        InputStream stream = MCRResourceHelper.getResourceAsStream(path);
-        if (stream != null) {
-            try (BufferedReader reader = new BufferedReader(new InputStreamReader(stream, UTF_8), 128)) {
-                return Optional.of(readUnmodifiableConfiguration(reader));
-            } catch (IOException e) {
-                throw new MCRException("Unable to read resource " + path);
-            }
-        }
-
-        return Optional.empty();
-
-    }
-
-    private static UnmodifiableConfiguration readUnmodifiableConfiguration(BufferedReader reader) throws IOException {
-        String className = reader.readLine();
-        String hint = reader.readLine();
-        return new UnmodifiableConfiguration(className, hint);
-    }
-
-    private UnmodifiableConfiguration toUnmodifiableConfiguration(MCRPasswordCheckStrategy strategy) {
-        return new UnmodifiableConfiguration(strategy.getClass().getName(), strategy.unmodifiableConfigurationHint());
-    }
-
-    private void checkConfigurationHasNoIncompatibleChange(String type, UnmodifiableConfiguration oldConfiguration,
-                                                           UnmodifiableConfiguration newConfiguration) {
-
-        if (!oldConfiguration.className().equals(newConfiguration.className())) {
-            throw new MCRConfigurationException("Detected incompatible implementation change for password " +
-                "check strategy " + type + " that will prevent existing password hashes from being " +
-                "successfully verified, even if the correct password was supplied, got " +
-                newConfiguration.className() + ", expected " + oldConfiguration.className());
-        }
-
-        if (!oldConfiguration.hint().equals(newConfiguration.hint())) {
-            throw new MCRConfigurationException("Detected incompatible value change for password " +
-                "check strategy " + type + " that will prevent existing password hashes from being " +
-                "successfully verified, even if the correct password was supplied, got " +
-                newConfiguration.hint() + ", expected " + oldConfiguration.hint());
-        }
-
-    }
-
-    private static void storeUnmodifiableConfiguration(String name, UnmodifiableConfiguration configuration) {
-
-        String path = DATA_DIRECTORY_NAME + "/" + name;
-
-        File file = MCRConfigurationDir.getConfigFile(path);
-        if (file != null) {
-
-            File parentDir = file.getParentFile();
-            if (!parentDir.exists()) {
-                boolean parentDirCreated = parentDir.mkdirs();
-                if (!parentDirCreated) {
-                    throw new MCRException("Unable to create value directory " + parentDir.getAbsolutePath());
-                }
-            }
-
-            try (BufferedWriter writer = new BufferedWriter(new FileWriter(file, UTF_8), 128)) {
-                writer.write(configuration.className());
-                writer.newLine();
-                writer.write(configuration.hint());
-                writer.newLine();
-            } catch (IOException e) {
-                throw new MCRException("Unable to write to value file " + file.getAbsolutePath());
-            }
-
-        }
-
     }
 
     public static MCRPasswordCheckManager instance() {
@@ -278,17 +147,6 @@ public final class MCRPasswordCheckManager {
         return strategy;
     }
 
-    private record UnmodifiableConfiguration(String className, String hint) {
-    }
-
-    public enum ConfigurationCheck {
-
-        OUTDATED_STRATEGY,
-
-        INCOMPATIBLE_CHANGE;
-
-    }
-
     public static class Factory implements Supplier<MCRPasswordCheckManager> {
 
         @MCRInstanceMap(name = STRATEGIES_KEY, valueClass = MCRPasswordCheckStrategy.class)
@@ -303,11 +161,12 @@ public final class MCRPasswordCheckManager {
         @Override
         public MCRPasswordCheckManager get() {
 
+            SecureRandom random = getStrongSecureRandom();
+
             Set<ConfigurationCheck> configurationChecks = splitValue(this.configurationChecks)
                 .map(ConfigurationCheck::valueOf).collect(Collectors.toSet());
 
-            return new MCRPasswordCheckManager(getStrongSecureRandom(), strategies, selectedStrategy,
-                configurationChecks);
+            return new MCRPasswordCheckManager(random, strategies, selectedStrategy, configurationChecks);
 
         }
 
@@ -318,6 +177,14 @@ public final class MCRPasswordCheckManager {
                 throw new MCRException("Failed to obtain strong secure random number generator", e);
             }
         }
+
+    }
+
+    public enum ConfigurationCheck {
+
+        OUTDATED_STRATEGY,
+
+        INCOMPATIBLE_CHANGE;
 
     }
