+++ mycore-user2/src/main/java/org/mycore/user2/hash/MCRPasswordCheckUtils.java
@@ -25,6 +25,10 @@ import javax.crypto.SecretKeyFactory;
 
 import org.mycore.common.config.MCRConfigurationException;
 
+
+/**
+ * Utility class providing functions commonly used by implementations of {@link MCRPasswordCheckStrategy}.
+ */
 public abstract class MCRPasswordCheckUtils {
     
     private MCRPasswordCheckUtils() {
